+++ CHANGES.rst
@@ -23,9 +23,10 @@ Unreleased
 
   .. FIXME: versionchanged on update_feeds() etc.
 
+* Add ``--scheduled`` flag to the ``update`` command. (:issue:`332`)
 * Group mutually-exclusive attributes of :class:`~.FeedUpdateIntent`
   into its :attr:`~.FeedUpdateIntent.value` union attribute. (:issue:`332`)
-  
+
 * Fix bug introduced in `version 3.12 <Version 3.12_>`_ causing an assertion error
   when there are multiple entries with the same id in the same feed,
   or when parallel :meth:`~Reader.update_feeds` calls add the same entry.